+++ expressions/operators/Multiplication.h
@@ -6,5 +6,5 @@ class Multiplication : public BinaryOperator
 {
 public:
     Multiplication(ExpressionPtr left, ExpressionPtr right);
-    double evaluate() const;
+    int evaluate() const;
 };
