+++ expressions/operators/Subtraction.h
@@ -6,5 +6,5 @@ class Subtraction : public BinaryOperator
 {
 public:
     Subtraction(ExpressionPtr left, ExpressionPtr right);
-    double evaluate() const;
+    int evaluate() const;
 };
